Module dbus_bytestream::dispatch

Routes incoming messages to registered handlers

Examples

Getting a callback from a method call

use std::cell::Cell;
use dbus_bytestream::connection::{Connection, MessageSender};
use dbus_bytestream::dispatch::{MessageDispatcher, MessageHandler};
use dbus_bytestream::dispatch::MethodRetVal::EmptyReply;
use dbus_bytestream::message;

// Simulated incoming method call
let mth_msg = message::create_method_call("dest", "path", "interface", "method");
let mth_called = Cell::new(false);

let mut conn = Connection::connect_system().unwrap();
let mut dis = MessageDispatcher::new();

// Register a boxed closure with a method call
dis.add_method("path".into(), "interface".into(), "method".into(),
                Box::new(|_sender, mth| {
                    mth_called.set(true);
                    assert_eq!(mth.path, "path");
                    assert_eq!(mth.interface, "interface");
                    assert_eq!(mth.member, "method");
                    Ok(EmptyReply)
                } ));

/* Handle one method call */
assert_eq!(mth_called.get(), false);
dis.handle_message(&mut conn, mth_msg).unwrap().unwrap();
assert_eq!(mth_called.get(), true);

Chaining multiple message handlers

use dbus_bytestream::connection::{Connection, MessageSender};
use dbus_bytestream::dispatch::{HandlerChain, MessageDispatcher,
                                MessageHandler, NoMatchHandler};
use dbus_bytestream::message;

let msg = message::create_method_call("dest", "path", "interface", "method");
let mut conn = Connection::connect_system().unwrap();
// Three dispatchers with no methods registered. Each will pass the msg to the next
let mut dis1 = MessageDispatcher::new();
let mut dis2 = MessageDispatcher::new();
let mut dis3 = MessageDispatcher::new();

// The NoMatchHandler will consume the msg and send an error over the connection.
dis1.handle_message(&mut conn, msg)
    .chain(&mut conn, &mut dis2)
    .chain(&mut conn, &mut dis3)
    .chain(&mut conn, &mut NoMatchHandler).unwrap().unwrap();

Modules

message_types
Specific struct representations of DBus message types

wrapper
Macros to generate message handlers from Rust impl blocks.

Structs

MessageDispatcher
Dispatches incoming signals and method calls

MessageSenderWrapper
Wrapper around a MessageSender trait object to work around a compiler bug

NoMatchHandler
A Handler that consumes a message and sends an error DBus message to the sender

Enums

DispatchError

MethodRetVal

Traits

HandlerChain
Allows chaining together message handlers

MessageHandler
Dispatches messages. Multiple message handlers can be chained.

Type Definitions

DispatchResult

HandlerResult

MethodHandler

MethodHandlerResult

SignalHandler

SignalHandlerResult