Trait dbus_bytestream::dispatch::MessageHandler

pub trait MessageHandler {
    fn handle_message<T: MessageSender>(&mut self, sender: &mut T, msg: Message) -> HandlerResult;
}

Dispatches messages. Multiple message handlers can be chained.

Required Methods

fn handle_message<T: MessageSender>(&mut self, sender: &mut T, msg: Message) -> HandlerResult

Dispatches one message

Handled messages are consumed and Ok(Ok(())) is returned. Invalid messages that fail to decode push a DBus error message to the sender and return Ok(Err(String)). Unhandled but valid messages are returned as Err(Message) to allow chaining dispatchers.