Module dbus_bytestream::dispatch::wrapper

Macros to generate method call and signal handlers from Rust impl blocks.

These macros bind Rust functions to generic closures which can be used with the message dispatcher.

For both signals and methods, the glue code validates the arguments in the message body received from the bus and will return error responses to the bus without invoking the callback.

Methods may optionally return a tuple of Encodables which will be inserted in the body of the autogenerated method call response message. The choice of a method to produce an autogenerated response is decided at run time via an Option return. This allows method calls to dynamically make the decision to defer the response message.

Method return arguments are guaranteed to be Encodable at compile time. Since the methods have a signature known at compile time, this also ensures a single method will always have the same return signature.

Examples

extern crate dbus_serialize;
#[macro_use]
extern crate dbus_bytestream;

use std::cell::Cell;

use dbus_bytestream::connection::MessageSender;
use dbus_bytestream::dispatch::wrapper::HandlerError;
use dbus_bytestream::dispatch::wrapper::SignalReturn;

#[derive(Default)]
struct MySignalHandler {
    val: Cell<i32>,
}

dbus_signal_impl! {
    impl MySignalHandler {
        fn TheSignalName(&self, _sender: &mut MessageSender, val: i32) -> SignalReturn {
            self.val.set(val);
            Ok(())
        }
    }
}

fn main() {
    use dbus_bytestream::connection::Connection;
    use dbus_bytestream::dispatch::{MessageDispatcher, MessageHandler};
    use dbus_bytestream::message;

    let mut conn = Connection::connect_system().unwrap();
    /* Create our signal handler object */
    let handler: MySignalHandler = Default::default();

    /* Register the signal handlers with a dispatcher */
    let mut dis = MessageDispatcher::new();
    for (name, handler) in handler.get_signal_handlers().into_iter() {
        dis.add_signal("path".into(), "interface".into(), name.into(), handler);
    }

    let sig = message::create_signal("path", "interface", "TheSignalName")
                .add_arg(&5);

    assert_eq!(handler.val.get(), 0);
    dis.handle_message(&mut conn, sig);
    assert_eq!(handler.val.get(), 5);
}

Type Definitions

HandlerError

MethodReturn
Return value for method handler functions

SignalReturn
Return value for signal handler functions