Module dbus_bytestream::message

Functions for creating and modifying messages to send across the message bus.

Structs

HeaderField

Message
Represents a received message from the message bus

MessageType

Constants

FLAGS_NO_REPLY_EXPECTED

HEADER_FIELD_DESTINATION

HEADER_FIELD_ERROR_NAME

HEADER_FIELD_INTERFACE

HEADER_FIELD_INVALID

HEADER_FIELD_MEMBER

HEADER_FIELD_PATH

HEADER_FIELD_REPLY_SERIAL

HEADER_FIELD_SENDER

HEADER_FIELD_SIGNATURE

MESSAGE_TYPE_ERROR

MESSAGE_TYPE_INVALID

MESSAGE_TYPE_METHOD_CALL

MESSAGE_TYPE_METHOD_RETURN

MESSAGE_TYPE_SIGNAL

Functions

create_error
Create a Message for a D-Bus error. Once created, return values can be added with Message.add_arg

create_method_call
Create a Message for a D-Bus method call. Once a Message is created, arguments can be added with Message.add_arg

create_method_return
Create a Message for a D-Bus method return. Once created, return values can be added with Message.add_arg

create_signal
Create a Message for a D-Bus signal. Once created, return values can be added with Message.add_arg