Struct dbus_bytestream::message::MessageBuf

pub struct MessageBuf(pub Vec<u8>);

Container for the header of a D-Bus message. Arguments can be added with add_arg()

Methods

impl MessageBuf

fn add_arg(self, arg: &Marshal) -> MessageBuf

Add the given argument to the MessageBuf. Accepts anything that implements the Marshal trait, which is most basic types, as well as the general-purpose dbus_serialize::types::Value enum.

Note that these calls can be chained together to add multiple arguments, see the example

Examples

dbus_bytestream::message::create_method_call("foo", "/bar", "baz", "bloop")
    .add_arg(&1)
    .add_arg(&"string");