Module dbus_bytestream::message

Functions for creating and modifying messages to send across the message bus.

Structs

Message
Represents a received message from the message bus

MessageBuf
Container for the header of a D-Bus message. Arguments can be added with add_arg()

MessageType

Enums

HeaderFieldName

Constants

MESSAGE_TYPE_ERROR

MESSAGE_TYPE_INVALID

MESSAGE_TYPE_METHOD_CALL

MESSAGE_TYPE_METHOD_RETURN

MESSAGE_TYPE_SIGNAL

Functions

create_method_call
Create a MessageBuf for a D-Bus method call. Once a MessageBuf object is created, arguments can be added to the object with MessageBuf.add_arg

get_length

set_length